+++ src/main/java/com/api/clinica/domain/service/UserService.java
@@ -3,6 +3,7 @@ package com.api.clinica.domain.service;
 import com.api.clinica.domain.data.entities.UserEntity;
 import com.api.clinica.domain.data.repositories.IUserRepository;
 import com.api.clinica.domain.dto.UserDTO;
+import com.api.clinica.domain.service.mappers.UserMapper;
 import com.api.clinica.infra.errors.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
