+++ src/main/java/com/api/clinica/domain/service/mappers/MedicoMapper.java
@@ -1,11 +1,11 @@
-package com.api.clinica.domain.service;
+package com.api.clinica.domain.service.mappers;
 
 import com.api.clinica.domain.data.entities.MedicoEntity;
 import com.api.clinica.domain.dto.DataMedicoDTO;
 import com.api.clinica.domain.dto.MedicoDTO;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
-@Service
+@Component
 public class MedicoMapper {
 
     public MedicoEntity convertMedicoDTOToMedicoEntity(MedicoDTO medicoDTO) {
